# swagger.yaml
---
openapi: 3.0.0
servers: []
info:
  description: This is a reference API
  version: "1.0.0"
  title: Reference API
paths:
  /comments:
    post:
      summary: Create a new comment
      operationId: createComment
      description: Create a new comment
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
        description: The new comment

  /articles:
    get:
      summary: Gets all articles.
      operationId: getArticles
      description: |
        Gets all articles.
      responses:
        '200':
          description: All the articles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleCollection'
    post:
      summary: Creates a new article
      operationId: createArticle
      description: It makes a new article authored by you, the current user
      responses:
        '201':
          description: Created article
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleRequest'
        description: The new article

  /articles/{id}:
    parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the article

    get:
      summary: Get a single article
      operationId: getArticle
      description: Gets a single article
      responses:
        '200':
          description: The article identified by the ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a single article
      operationId: updateArticle
      description: Updates a single article
      responses:
        '204':
          description: Updated article successfully
        '403':
          description: You are not authorized to update this article
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleRequest'
    delete:
      summary: Delete a single article
      operationId: deletePost
      description: Deletes a single article
      responses:
        '200':
          description: The deleted resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '403':
          description: You are not authorized to delete this article
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /articles/{id}/comments:
    get:
      summary: Get all comments for an article
      operationId: getArticleComments
      description: Get all comments for an article
      responses:
        '200':
          description: Comments on the article
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentCollection'
  /users/current:
    get:
      summary: Get the current user
      operationId: getCurrentUser
      description: Gets the current user
      responses:
        '200':
          description: The current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
components:
  schemas:
    User:
      type: object
      properties:
        name:
          type: string
          example: Joan Baez
    Article:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 12345-abcdef-qwerty-foobar
        title:
          type: string
          example: Goto Considered Harmful
        body:
          type: string
          example: |
            To be or not to be, that is the question
        author:
          $ref: '#/components/schemas/User'
    ArticleRequest:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
    ArticleCollection:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: '#/components/schemas/Article'
    Comment:
      type: object
      properties:
        body:
          type: string
          example: great job!
        id:
          type: string
        articleId:
          type: string
          format: uuid
    CommentRequest:
      type: object
      properties:
        body:
          type: string
          example: love this article
        articleId:
          type: string
          example: 123456-abcdef-qwerty-foobar
    CommentCollection:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
    Error:
      type: object
      properties:
        message:
          type: string
          example: Oops! That is not a valid id
        error:
          type: string
          example: Forbidden
        path:
          type: string
          example: /v1/articles